backend/requirements.txt
asgiref==3.7.2
black==23.12.1
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
